Add BtSpaces Databag and BtSpaceMarker component
The `BtSpaces` is a `Databag` that stores the data of all the physical
worlds. It has at max 4 physical worlds (called Spaces).

The `BtSpaceMarker` is an optional component that can be used to specify to which
`Space` the `Entity` belongs. When this component is not specified, the
`Entity` is put to the main `Space`.